Use image tag "latest" for prov-app
As the deployment happens via Jenkins, the "latest" tag is set.

While we still have the prov-dev/prov-test setup, this is the best
configuration. If and when there is a prov-app installation in the "cd"
namespace, then this should probably be tagged with ODS_IMAGE_TAG. But
will do that then ...

Fixes #331.